feat: add lib support for rn new architecture on android
