+++ android/gradle.properties
@@ -1,4 +1,4 @@
-MlkitOcr_kotlinVersion=1.3.50
-MlkitOcr_compileSdkVersion=28
-MlkitOcr_buildToolsVersion=28.0.3
-MlkitOcr_targetSdkVersion=28
+MlkitOcr_kotlinVersion=1.6.10
+MlkitOcr_minSdkVersion=21
+MlkitOcr_compileSdkVersion=31
+MlkitOcr_targetSdkVersion=31
